Updated Loot System to use EntityBase
BreakableBase is now defunct, replaced with functionality from EntityBase. LootContainer requires an EntityBase companion, and activates off of self-referential UnityEvent. Made changes to default projectile script and playerController to work better with EntityBase.

+++ Assets/_Game/Scripts/Mechanics/Player/Weapons&Aiming/WeaponBase.cs
@@ -17,7 +17,7 @@ public class WeaponBase : MonoBehaviour
 	[Header("Fire Settings")]
 	[SerializeField] float clickCooldown = 0.5f;
 	private float cdTime = 0f;
-	[SerializeField] float damage = 3f;
+	[SerializeField] int damage = 3;
 	[SerializeField] float projectileSpeed = 40f;
 	[SerializeField] [Range(0, 45)] float projectileCone = 6f;
